requirements.txt
python-telegram-bot
simplejson
beautifulsoup4
dbus-python
SQLAlchemy
